add styling to Login and register pages
