+++ frontend/src/App.jsx
@@ -5,8 +5,8 @@ import ImageGeneration from "./components/ImageGeneration";
 import ImageEditing from "./components/ImageEditing";
 import Creations from "./components/Creations";
 import SliderPage from "./components/SliderPage";
-import Login from "./components/Login";
-import Register from "./components/Register";
+import Login from "./components/Login/Login";
+import Register from "./components/Register/Register";
 import { AuthProvider } from "./context/AuthContext";
 import Profile from "./components/Profile";
 import ProtectedRoute from "./components/ProtectedRoute";
