+++ frontend/src/components/Register/Register.jsx
@@ -1,18 +1,15 @@
 import { useState } from "react";
 import { Link, useNavigate } from "react-router-dom";
-import {
-  signInWithEmail,
-  signInWithGoogle,
-  resetPassword,
-} from "../services/auth";
+import { signUpWithEmail, signInWithGoogle } from "../../services/auth";
 
-const Login = () => {
+const Register = () => {
   const [loading, setLoading] = useState(false);
   const [error, setError] = useState(null);
   const [success, setSuccess] = useState(null);
   const [formData, setFormData] = useState({
     email: "",
     password: "",
+    displayName: "",
   });
   const navigate = useNavigate();
 
@@ -28,29 +25,36 @@ const Login = () => {
     setSuccess(null);
 
     try {
-      // Handle login
-      const result = await signInWithEmail(formData.email, formData.password);
+      // Handle registration
+      const user = await signUpWithEmail(
+        formData.email,
+        formData.password,
+        formData.displayName
+      );
+
+      // For registration, we need to get a token first
+      const idToken = await user.getIdToken();
 
       // Store auth user and token in localStorage
       localStorage.setItem(
         "authUser",
         JSON.stringify({
-          uid: result.user.uid,
-          email: result.user.email,
-          displayName: result.user.displayName,
-          token: result.token,
+          uid: user.uid,
+          email: user.email,
+          displayName: user.displayName,
+          token: idToken,
         })
       );
 
-      setSuccess("Login successful!");
+      setSuccess("Registration successful!");
 
       // Clear form
-      setFormData({ email: "", password: "" });
+      setFormData({ email: "", password: "", displayName: "" });
 
-      // Redirect to home or intended page
+      // Redirect to home or login page
       setTimeout(() => navigate("/"), 1000);
     } catch (err) {
-      setError(err.message || "Authentication failed");
+      setError(err.message || "Registration failed");
     } finally {
       setLoading(false);
     }
@@ -75,9 +79,9 @@ const Login = () => {
         })
       );
 
-      setSuccess("Login with Google successful!");
+      setSuccess("Registration with Google successful!");
 
-      // Redirect to home or intended page
+      // Redirect to home
       setTimeout(() => navigate("/"), 1000);
     } catch (err) {
       setError(err.message || "Google authentication failed");
@@ -86,34 +90,14 @@ const Login = () => {
     }
   };
 
-  const handlePasswordReset = async () => {
-    if (!formData.email) {
-      setError("Please enter your email address");
-      return;
-    }
-
-    setLoading(true);
-    setError(null);
-    setSuccess(null);
-
-    try {
-      await resetPassword(formData.email);
-      setSuccess("Password reset email sent. Please check your inbox.");
-    } catch (err) {
-      setError(err.message || "Failed to send password reset email");
-    } finally {
-      setLoading(false);
-    }
-  };
-
   return (
-    <div className="auth-container">
-      <h2>Login</h2>
+    <div className="register-container">
+      <h2>Register</h2>
 
       {error && <div className="error-message">{error}</div>}
       {success && <div className="success-message">{success}</div>}
 
-      <form onSubmit={handleSubmit}>
+      <form onSubmit={handleSubmit} className="register-form">
         <div className="form-group">
           <label htmlFor="email">Email</label>
           <input
@@ -136,42 +120,56 @@ const Login = () => {
             onChange={handleInputChange}
             required
           />
+          <div className="input-hint">
+            Password must be at least 6 characters long
+          </div>
+        </div>
+
+        <div className="form-group">
+          <label htmlFor="displayName">Display Name</label>
+          <input
+            type="text"
+            id="displayName"
+            name="displayName"
+            value={formData.displayName}
+            onChange={handleInputChange}
+          />
         </div>
 
         <div className="form-actions">
-          <button type="submit" disabled={loading}>
-            {loading ? "Processing..." : "Login"}
-          </button>
-
-          <button
-            type="button"
-            onClick={handlePasswordReset}
-            disabled={loading}
-            className="reset-button"
-          >
-            Forgot Password?
+          <button type="submit" disabled={loading} className="register-button">
+            {loading ? "Processing..." : "Register"}
           </button>
         </div>
       </form>
 
+      <div className="terms">
+        By registering, you agree to our <a href="#">Terms of Service</a> and{" "}
+        <a href="#">Privacy Policy</a>
+      </div>
+
       <div className="social-login">
+        <div className="divider">
+          <span>OR</span>
+        </div>
         <button
           type="button"
           onClick={handleGoogleSignIn}
           disabled={loading}
           className="google-button"
         >
-          Continue with Google
+          <img src="/google-icon.svg" alt="Google" className="google-icon" />
+          Register with Google
         </button>
       </div>
 
       <div className="auth-toggle">
-        <Link to="/register" className="toggle-link">
-          Don't have an account? Register
+        <Link to="/login" className="toggle-link">
+          Already have an account? Login
         </Link>
       </div>
     </div>
   );
 };
 
-export default Login;
+export default Register;
