+++ frontend/src/components/Register/_register.scss
@@ -1,0 +1,173 @@
+@import "../../variables.scss";
+.register-container {
+  max-width: 450px;
+  margin: 2rem auto;
+  padding: 2rem;
+  background-color: $card-bg;
+  border-radius: 8px;
+  box-shadow: 0 4px 20px rgba(0, 0, 0, 0.3);
+  border: 1px solid $border-color;
+  transition: all 0.3s ease;
+
+  h2 {
+    color: $accent-red;
+    text-align: center;
+    margin-bottom: 1.5rem;
+    font-size: 2rem;
+    font-weight: 600;
+    border-bottom: 1px solid $border-color;
+    padding-bottom: 1rem;
+  }
+
+  .register-form {
+    .form-group {
+      margin-bottom: 1.5rem;
+
+      label {
+        display: block;
+        margin-bottom: 0.5rem;
+        font-weight: 500;
+        color: $text-light;
+      }
+
+      input {
+        width: 100%;
+        padding: 12px 16px;
+        font-size: 1rem;
+        border: 1px solid $border-color;
+        border-radius: 6px;
+        background-color: $input-bg;
+        color: $text-light;
+        transition: border-color 0.2s ease;
+
+        &:focus {
+          outline: none;
+          border-color: $accent-red;
+          box-shadow: 0 0 0 2px rgba($accent-red, 0.2);
+        }
+      }
+
+      .input-hint {
+        font-size: 0.8rem;
+        color: rgba($text-light, 0.7);
+        margin-top: 0.5rem;
+      }
+    }
+
+    .form-actions {
+      display: flex;
+      flex-direction: column;
+      gap: 1rem;
+      margin-top: 2rem;
+
+      .register-button {
+        background-color: $accent-red;
+        color: white;
+        padding: 12px;
+        border: none;
+        border-radius: 6px;
+        font-size: 1rem;
+        font-weight: 600;
+        cursor: pointer;
+        transition: background-color 0.2s ease;
+
+        &:hover {
+          background-color: darken($accent-red, 10%);
+        }
+
+        &:disabled {
+          opacity: 0.7;
+          cursor: not-allowed;
+        }
+      }
+    }
+  }
+
+  .terms {
+    font-size: 0.8rem;
+    color: rgba($text-light, 0.7);
+    text-align: center;
+    margin-top: 1rem;
+
+    a {
+      color: $accent-red;
+      text-decoration: none;
+
+      &:hover {
+        text-decoration: underline;
+      }
+    }
+  }
+
+  .social-login {
+    margin-top: 2rem;
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+
+    .divider {
+      display: flex;
+      align-items: center;
+      width: 100%;
+      margin: 1.5rem 0;
+
+      &::before,
+      &::after {
+        content: "";
+        flex: 1;
+        height: 1px;
+        background-color: $border-color;
+      }
+
+      span {
+        padding: 0 1rem;
+        color: $text-light;
+        font-size: 0.9rem;
+      }
+    }
+
+    .google-button {
+      display: flex;
+      align-items: center;
+      justify-content: center;
+      gap: 0.5rem;
+      width: 100%;
+      padding: 12px;
+      background-color: white;
+      color: #444;
+      border: 1px solid #ddd;
+      border-radius: 6px;
+      font-size: 1rem;
+      font-weight: 500;
+      cursor: pointer;
+      transition: background-color 0.2s ease;
+
+      &:hover {
+        background-color: #f5f5f5;
+      }
+
+      .google-icon {
+        width: 18px;
+        height: 18px;
+      }
+    }
+  }
+
+  .auth-toggle {
+    margin-top: 2rem;
+    text-align: center;
+
+    .toggle-link {
+      color: $text-light;
+      text-decoration: none;
+      font-size: 0.9rem;
+
+      &:hover {
+        color: $accent-red;
+        text-decoration: underline;
+      }
+    }
+  }
+}
+
+// Error and success messages are shared with Login.scss
