+++ frontend/src/styles.scss
@@ -1,14 +1,6 @@
-// Main variables
-$main-bg: #1b1a1f;
-$accent-red: #b5332f;
-$text-light: #f5f5f5;
-$text-dark: #333;
-$border-color: #444;
-$input-bg: #444;
-$card-bg: #222;
-$dark-card-bg: #2a2a2a;
-$success-green: #10b981;
-$secondary-blue: #0077cc;
+@import "./components/Login/_login.scss";
+@import "./components/Register/_register.scss";
+@import "./variables.scss";
 
 // Reset margin and padding on html and body
 html,
